--- Projects/App/.swiftlint.yml
@@ -1,20 +1,0 @@
-excluded:
-  - "**/*/NeedleGenerated.swift"
-  - "Tuist"
-
-
-force_cast:     # From https://realm.github.io/SwiftLint/force_cast.html
-    warning     # 강제 캐스팅은 error에서 warning으로 변경
-
-disabled_rules:
-  - trailing_whitespace
-  - vertical_whitespace
-  - colon
-  - comma
-  - comment_spacing # 주석 여백
-  - mark # //MARK 관련 warning
-  - opening_brace # 괄호 열 때 줄바꿈을 했는지 안했는지, (안했으면 warning)
-  - line_length # 한줄에 글자 수 제한 , warning: 120, error: 200 (default) 
-  - statement_position # if는 반드시 else로 끝나야한다.
-  - identifier_name #변수명 글자 수 제한
-
