--- Projects/Features/CommonFeature/Resources/CommonFeature.txt
@@ -1,1 +1,0 @@
-// This is for tuist
