+++ Projects/Features/CommonFeature/Resources/RecommendPlayList/RecommendPlayListView.xib
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21507" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_12" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="RecommendPlayListView" customModule="CommonFeature" customModuleProvider="target">
+            <connections>
+                <outlet property="collectionView" destination="LaC-EM-sWn" id="J7F-eI-0te"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="Hek-wG-Vw9">
+            <rect key="frame" x="0.0" y="0.0" width="393" height="200"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="왁뮤팀이 추천하는 플레이리스트" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="KgG-tE-guY">
+                    <rect key="frame" x="20" y="32" width="353" height="24"/>
+                    <constraints>
+                        <constraint firstAttribute="height" constant="24" id="XBT-GN-yg1"/>
+                    </constraints>
+                    <fontDescription key="fontDescription" type="boldSystem" pointSize="16"/>
+                    <nil key="textColor"/>
+                    <nil key="highlightedColor"/>
+                </label>
+                <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="LaC-EM-sWn">
+                    <rect key="frame" x="20" y="76" width="353" height="104"/>
+                    <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                    <collectionViewFlowLayout key="collectionViewLayout" minimumLineSpacing="10" minimumInteritemSpacing="10" id="jem-Sq-3xE">
+                        <size key="itemSize" width="128" height="128"/>
+                        <size key="headerReferenceSize" width="0.0" height="0.0"/>
+                        <size key="footerReferenceSize" width="0.0" height="0.0"/>
+                        <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
+                    </collectionViewFlowLayout>
+                </collectionView>
+            </subviews>
+            <viewLayoutGuide key="safeArea" id="FSS-Vl-StI"/>
+            <color key="backgroundColor" red="0.94117647059999998" green="0.95294117649999999" blue="0.96470588239999999" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
+            <constraints>
+                <constraint firstItem="KgG-tE-guY" firstAttribute="top" secondItem="Hek-wG-Vw9" secondAttribute="top" constant="32" id="1l4-oJ-5I6"/>
+                <constraint firstItem="FSS-Vl-StI" firstAttribute="trailing" secondItem="LaC-EM-sWn" secondAttribute="trailing" constant="20" id="LXU-7q-Qex"/>
+                <constraint firstItem="FSS-Vl-StI" firstAttribute="bottom" secondItem="LaC-EM-sWn" secondAttribute="bottom" constant="20" id="UGh-up-S52"/>
+                <constraint firstItem="LaC-EM-sWn" firstAttribute="top" secondItem="KgG-tE-guY" secondAttribute="bottom" constant="20" id="YmR-JK-nbd"/>
+                <constraint firstItem="KgG-tE-guY" firstAttribute="leading" secondItem="FSS-Vl-StI" secondAttribute="leading" constant="20" id="dCj-rg-OV0"/>
+                <constraint firstItem="FSS-Vl-StI" firstAttribute="trailing" secondItem="KgG-tE-guY" secondAttribute="trailing" constant="20" id="o0j-pP-sJe"/>
+                <constraint firstItem="LaC-EM-sWn" firstAttribute="leading" secondItem="FSS-Vl-StI" secondAttribute="leading" constant="20" id="oFJ-WY-Fyf"/>
+            </constraints>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
+            <point key="canvasLocation" x="138" y="21"/>
+        </view>
+    </objects>
+</document>
