+++ Projects/Features/CommonFeature/Resources/SongList/SongListCell.xib
@@ -1,0 +1,79 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21507" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_12" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="SongListCell" customModule="CommonFeature" customModuleProvider="target">
+            <connections>
+                <outlet property="albumImageView" destination="uuH-Da-aO8" id="XsF-32-XsA"/>
+                <outlet property="artistLabel" destination="puI-a7-gq3" id="0PF-Nx-2va"/>
+                <outlet property="releaseDateLabel" destination="hyN-WV-WEd" id="3WL-08-Kck"/>
+                <outlet property="titleLabel" destination="yCz-Z2-lxg" id="ovp-A6-J2y"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="default" indentationWidth="10" id="JQV-bU-fkN" customClass="SongListCell" customModule="DesignSystem" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="390" height="44"/>
+            <autoresizingMask key="autoresizingMask"/>
+            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" tableViewCell="JQV-bU-fkN" id="Vwu-DM-hzC">
+                <rect key="frame" x="0.0" y="0.0" width="390" height="44"/>
+                <autoresizingMask key="autoresizingMask"/>
+                <subviews>
+                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="dummy_thumbnail_large" translatesAutoresizingMaskIntoConstraints="NO" id="uuH-Da-aO8">
+                        <rect key="frame" x="20" y="2" width="72" height="40"/>
+                        <constraints>
+                            <constraint firstAttribute="height" constant="40" id="kES-oJ-Dhl"/>
+                            <constraint firstAttribute="width" constant="72" id="mcr-vB-Yc1"/>
+                        </constraints>
+                    </imageView>
+                    <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" alignment="top" translatesAutoresizingMaskIntoConstraints="NO" id="fvW-ao-FDV">
+                        <rect key="frame" x="99.999999999999986" y="1.6666666666666679" width="208.66666666666663" height="40.666666666666657"/>
+                        <subviews>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="yCz-Z2-lxg">
+                                <rect key="frame" x="0.0" y="0.0" width="41.333333333333336" height="20.333333333333332"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                                <nil key="textColor"/>
+                                <nil key="highlightedColor"/>
+                            </label>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="puI-a7-gq3">
+                                <rect key="frame" x="0.0" y="20.333333333333336" width="41.333333333333336" height="20.333333333333336"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                                <nil key="textColor"/>
+                                <nil key="highlightedColor"/>
+                            </label>
+                        </subviews>
+                    </stackView>
+                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="hyN-WV-WEd">
+                        <rect key="frame" x="328.66666666666669" y="11.999999999999998" width="41.333333333333314" height="20.333333333333329"/>
+                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                        <nil key="textColor"/>
+                        <nil key="highlightedColor"/>
+                    </label>
+                </subviews>
+                <constraints>
+                    <constraint firstItem="fvW-ao-FDV" firstAttribute="centerY" secondItem="uuH-Da-aO8" secondAttribute="centerY" id="3pW-2b-bSR"/>
+                    <constraint firstAttribute="trailing" secondItem="hyN-WV-WEd" secondAttribute="trailing" constant="20" id="Bxj-SL-NSJ"/>
+                    <constraint firstItem="uuH-Da-aO8" firstAttribute="centerY" secondItem="Vwu-DM-hzC" secondAttribute="centerY" id="Ok7-o6-GNh"/>
+                    <constraint firstItem="fvW-ao-FDV" firstAttribute="leading" secondItem="uuH-Da-aO8" secondAttribute="trailing" constant="8" id="QWZ-jj-QJr"/>
+                    <constraint firstItem="hyN-WV-WEd" firstAttribute="centerY" secondItem="Vwu-DM-hzC" secondAttribute="centerY" id="Z3X-WH-ro0"/>
+                    <constraint firstItem="hyN-WV-WEd" firstAttribute="leading" secondItem="fvW-ao-FDV" secondAttribute="trailing" constant="20" id="ZVv-7N-bcJ"/>
+                    <constraint firstItem="uuH-Da-aO8" firstAttribute="leading" secondItem="Vwu-DM-hzC" secondAttribute="leading" constant="20" id="qh6-i3-Pds"/>
+                </constraints>
+            </tableViewCellContentView>
+            <connections>
+                <outlet property="albumImageView" destination="uuH-Da-aO8" id="eht-pB-cgp"/>
+                <outlet property="artistLabel" destination="puI-a7-gq3" id="8Bm-4h-Skr"/>
+                <outlet property="releaseDateLabel" destination="hyN-WV-WEd" id="5tD-wz-X4U"/>
+                <outlet property="titleLabel" destination="yCz-Z2-lxg" id="GmU-81-at9"/>
+            </connections>
+            <point key="canvasLocation" x="185" y="-163"/>
+        </tableViewCell>
+    </objects>
+    <resources>
+        <image name="dummy_thumbnail_large" width="324" height="182"/>
+    </resources>
+</document>
