+++ Projects/Features/CommonFeature/Resources/Warning/Warning.xib
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21507" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_12" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="System colors in document resources" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="WarningView" customModule="CommonFeature" customModuleProvider="target">
+            <connections>
+                <outlet property="superView" destination="WIK-TM-FxM" id="0lW-d2-9Pv"/>
+                <outlet property="warningImageView" destination="M3b-Bu-yMi" id="mK7-jM-9qc"/>
+                <outlet property="warningLabelView" destination="zfG-ec-HaH" id="zda-Zv-QMz"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="28g-6R-llj">
+            <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="WIK-TM-FxM">
+                    <rect key="frame" x="176" y="375.66666666666669" width="41.333333333333343" height="101"/>
+                    <subviews>
+                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="warning" translatesAutoresizingMaskIntoConstraints="NO" id="M3b-Bu-yMi">
+                            <rect key="frame" x="-19.333333333333343" y="0.0" width="80" height="80"/>
+                        </imageView>
+                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="zfG-ec-HaH">
+                            <rect key="frame" x="0.0" y="80" width="41.333333333333336" height="21"/>
+                            <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                            <nil key="textColor"/>
+                            <nil key="highlightedColor"/>
+                        </label>
+                    </subviews>
+                    <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                    <constraints>
+                        <constraint firstItem="zfG-ec-HaH" firstAttribute="top" secondItem="M3b-Bu-yMi" secondAttribute="bottom" id="GfK-1b-F0g"/>
+                        <constraint firstItem="M3b-Bu-yMi" firstAttribute="top" secondItem="WIK-TM-FxM" secondAttribute="top" id="UZE-LA-O2G"/>
+                        <constraint firstAttribute="trailing" secondItem="zfG-ec-HaH" secondAttribute="trailing" id="b2g-zs-GS5"/>
+                        <constraint firstItem="zfG-ec-HaH" firstAttribute="leading" secondItem="WIK-TM-FxM" secondAttribute="leading" id="bbk-qF-uGJ"/>
+                        <constraint firstAttribute="bottom" secondItem="zfG-ec-HaH" secondAttribute="bottom" id="ogd-ID-QvT"/>
+                        <constraint firstItem="M3b-Bu-yMi" firstAttribute="centerX" secondItem="WIK-TM-FxM" secondAttribute="centerX" id="rn4-S9-zRF"/>
+                    </constraints>
+                </view>
+            </subviews>
+            <viewLayoutGuide key="safeArea" id="2Z5-Yt-cCd"/>
+            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+            <constraints>
+                <constraint firstItem="WIK-TM-FxM" firstAttribute="centerX" secondItem="28g-6R-llj" secondAttribute="centerX" id="Yal-0K-xrv"/>
+                <constraint firstItem="WIK-TM-FxM" firstAttribute="centerY" secondItem="28g-6R-llj" secondAttribute="centerY" id="xfE-Ec-OeV"/>
+            </constraints>
+            <point key="canvasLocation" x="178" y="36"/>
+        </view>
+    </objects>
+    <resources>
+        <image name="warning" width="80" height="80"/>
+        <systemColor name="systemBackgroundColor">
+            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+        </systemColor>
+    </resources>
+</document>
