+++ Projects/Features/CommonFeature/Sources/Views/PlayListTableViewCell.swift
@@ -1,39 +1,36 @@
 //
-//  FavoriteTableViewCell.swift
+//  PlayListTableViewCell.swift
 //  StorageFeature
 //
-//  Created by yongbeomkwak on 2023/01/27.
+//  Created by yongbeomkwak on 2023/01/19.
 //  Copyright © 2023 yongbeomkwak. All rights reserved.
 //
 
 import UIKit
 import DesignSystem
-import CommonFeature
 
-class FavoriteTableViewCell: UITableViewCell {
-
-    @IBOutlet weak var albumImageView: UIImageView!
+class PlayListTableViewCell: UITableViewCell {
+    @IBOutlet weak var button:UIButton!
     
     @IBOutlet weak var titleLabel: UILabel!
+    
+    @IBOutlet weak var albumImageView: UIImageView!
     @IBOutlet weak var artistLabel: UILabel!
     
-    @IBOutlet weak var button: UIButton!
-    
-    var isEdit:Bool = false
-    
-    
-    
-
-    @IBAction func playAxtion(_ sender: UIButton) {
+    @IBAction func playOrEditAction(_ sender: UIButton) {
         
-        if !isEdit{
+        if isEdit == false
+        {
             print("Play")
         }
         
     }
+    var isEdit:Bool = false
+    
+    
     override func awakeFromNib() {
         super.awakeFromNib()
-        // Initialization code
+        
         self.backgroundColor = .clear
         
         albumImageView.layer.cornerRadius = 8
@@ -46,22 +43,30 @@ class FavoriteTableViewCell: UITableViewCell {
         
         self.titleLabel.textColor = DesignSystemAsset.GrayColor.gray900.color
         self.artistLabel.textColor = DesignSystemAsset.GrayColor.gray900.color
+
         
-        
+
+       
     }
+    
+    
+
+   
 
 }
 
+extension PlayListTableViewCell {
+    func update(_ model: SongInfoDTO,_ isEditing:Bool) {
 
-extension FavoriteTableViewCell {
-    
-    public func update(_ model:SongInfoDTO,_ isEditing:Bool)
-    {
+       
         albumImageView.image = DesignSystemAsset.Player.dummyThumbnailLarge.image
         titleLabel.text =  model.name
         artistLabel.text = model.artist
         isEdit = isEditing
         
         button.setImage( isEditing ? DesignSystemAsset.Storage.move.image :  DesignSystemAsset.Storage.play.image, for: .normal)
+        
+
+        
     }
 }
