+++ Projects/Features/HomeFeature/Sources/HomeViewController.swift
@@ -1,4 +1,5 @@
 import UIKit
+import CommonFeature
 import Utility
 import DesignSystem
 import PanModal
