+++ Projects/Features/SearchFeature/Sources/ViewControllers/SearchViewController.swift
@@ -1,4 +1,5 @@
 import UIKit
+import CommonFeature
 import DomainModule
 import NeedleFoundation
 import Utility
