+++ Projects/Features/SignInFeature/Sources/ViewControllers/LoginViewController.swift
@@ -13,6 +13,7 @@ import NaverThirdPartyLogin
 import RxSwift
 import Alamofire
 import AuthenticationServices
+import CommonFeature
 
 
 public class LoginViewController: UIViewController, ViewControllerFromStoryBoard {
