+++ Projects/Features/StorageFeature/Sources/ViewControllers/RequestViewController.swift
@@ -9,6 +9,7 @@
 import UIKit
 import Utility
 import DesignSystem
+import CommonFeature
 
 class RequestViewController: UIViewController, ViewControllerFromStoryBoard {
