--- Projects/UsertInterfaces/DesignSystem/Resources/PlayButtonGroup/PlayButtonGroupView.xib
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21225" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_0" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21207"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="System colors in document resources" minToolsVersion="11.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="PlayButtonGroupView" customModule="DesignSystem" customModuleProvider="target">
-            <connections>
-                <outlet property="allPlayButton" destination="f7I-eD-s9E" id="HPb-lk-8cz"/>
-                <outlet property="shufflePlayButton" destination="50f-Sn-UVp" id="Reb-GA-Qdw"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB">
-            <rect key="frame" x="0.0" y="0.0" width="500" height="162"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="hsP-wA-LdH">
-                    <rect key="frame" x="0.0" y="0.0" width="500" height="162"/>
-                    <subviews>
-                        <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="f7I-eD-s9E">
-                            <rect key="frame" x="20" y="16" width="226" height="52"/>
-                            <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                            <constraints>
-                                <constraint firstAttribute="height" constant="52" id="rzt-kp-cHE"/>
-                            </constraints>
-                            <fontDescription key="fontDescription" type="system" weight="medium" pointSize="14"/>
-                            <inset key="imageEdgeInsets" minX="0.0" minY="0.0" maxX="2.2250738585072014e-308" maxY="0.0"/>
-                            <state key="normal" title="전체재생">
-                                <color key="titleColor" red="0.062745098040000002" green="0.094117647060000004" blue="0.15686274510000001" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                            </state>
-                            <connections>
-                                <action selector="pressAllPlay:" destination="-1" eventType="touchUpInside" id="hBa-M5-LxQ"/>
-                            </connections>
-                        </button>
-                        <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="50f-Sn-UVp">
-                            <rect key="frame" x="254" y="16" width="226" height="52"/>
-                            <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                            <fontDescription key="fontDescription" type="system" weight="medium" pointSize="14"/>
-                            <inset key="imageEdgeInsets" minX="0.0" minY="0.0" maxX="2.2250738585072014e-308" maxY="0.0"/>
-                            <state key="normal" title="랜덤재생">
-                                <color key="titleColor" red="0.062745098040000002" green="0.094117647060000004" blue="0.15686274510000001" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-                            </state>
-                            <connections>
-                                <action selector="pressSufflePlay:" destination="-1" eventType="touchUpInside" id="s54-1U-tkJ"/>
-                            </connections>
-                        </button>
-                    </subviews>
-                    <constraints>
-                        <constraint firstItem="50f-Sn-UVp" firstAttribute="leading" secondItem="f7I-eD-s9E" secondAttribute="trailing" constant="8" id="05p-4L-d5q"/>
-                        <constraint firstAttribute="trailing" secondItem="50f-Sn-UVp" secondAttribute="trailing" constant="20" id="PFZ-QS-gec"/>
-                        <constraint firstItem="f7I-eD-s9E" firstAttribute="top" secondItem="hsP-wA-LdH" secondAttribute="top" constant="16" id="SH1-D9-94u"/>
-                        <constraint firstItem="f7I-eD-s9E" firstAttribute="height" secondItem="50f-Sn-UVp" secondAttribute="height" id="T2V-Cw-srV"/>
-                        <constraint firstItem="50f-Sn-UVp" firstAttribute="centerY" secondItem="f7I-eD-s9E" secondAttribute="centerY" id="lQi-Bv-CZQ"/>
-                        <constraint firstItem="f7I-eD-s9E" firstAttribute="width" secondItem="50f-Sn-UVp" secondAttribute="width" id="wBo-5D-Fq5"/>
-                        <constraint firstItem="f7I-eD-s9E" firstAttribute="leading" secondItem="hsP-wA-LdH" secondAttribute="leading" constant="20" id="wY3-hU-FAg"/>
-                    </constraints>
-                </view>
-            </subviews>
-            <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
-            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-            <constraints>
-                <constraint firstItem="hsP-wA-LdH" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" id="QT8-NY-Nll"/>
-                <constraint firstAttribute="top" secondItem="hsP-wA-LdH" secondAttribute="top" id="aJC-cU-ZaJ"/>
-                <constraint firstAttribute="trailing" secondItem="hsP-wA-LdH" secondAttribute="trailing" id="feO-LA-U2s"/>
-                <constraint firstItem="hsP-wA-LdH" firstAttribute="bottom" secondItem="iN0-l3-epB" secondAttribute="bottom" id="uxn-U7-SBr"/>
-            </constraints>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <point key="canvasLocation" x="172.30769230769229" y="66.824644549763036"/>
-        </view>
-    </objects>
-    <resources>
-        <systemColor name="systemBackgroundColor">
-            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-        </systemColor>
-    </resources>
-</document>
