--- Projects/UsertInterfaces/DesignSystem/Resources/RecommendPlayList/RecommendPlayListCell.xib
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21507" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_12" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
-        <capability name="collection view cell content view" minToolsVersion="11.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" reuseIdentifier="RecommendPlayListCell" id="D7k-Nl-NCm" customClass="RecommendPlayListCell" customModule="DesignSystem" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="164" height="80"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <collectionViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" insetsLayoutMarginsFromSafeArea="NO" id="Xxm-Vz-9PF">
-                <rect key="frame" x="0.0" y="0.0" width="164" height="80"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" translatesAutoresizingMaskIntoConstraints="NO" id="RHB-l7-KpO">
-                        <rect key="frame" x="100" y="16" width="48" height="48"/>
-                        <constraints>
-                            <constraint firstAttribute="width" secondItem="RHB-l7-KpO" secondAttribute="height" multiplier="1:1" id="fef-YH-GOn"/>
-                        </constraints>
-                    </imageView>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="캐롤" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="2" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="a81-Al-Kbi">
-                        <rect key="frame" x="16" y="31.666666666666671" width="80" height="17"/>
-                        <fontDescription key="fontDescription" type="system" weight="medium" pointSize="14"/>
-                        <color key="textColor" red="0.2784313725" green="0.3294117647" blue="0.4039215686" alpha="1" colorSpace="custom" customColorSpace="calibratedRGB"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                </subviews>
-                <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                <constraints>
-                    <constraint firstItem="RHB-l7-KpO" firstAttribute="leading" secondItem="a81-Al-Kbi" secondAttribute="trailing" constant="4" id="NgH-yx-z3X"/>
-                    <constraint firstAttribute="trailing" secondItem="RHB-l7-KpO" secondAttribute="trailing" constant="16" id="Plg-mx-3b6"/>
-                    <constraint firstItem="RHB-l7-KpO" firstAttribute="centerY" secondItem="Xxm-Vz-9PF" secondAttribute="centerY" id="XN6-nq-xAm"/>
-                    <constraint firstItem="RHB-l7-KpO" firstAttribute="height" secondItem="Xxm-Vz-9PF" secondAttribute="height" multiplier="48:80" id="cy9-9P-ePv"/>
-                    <constraint firstItem="a81-Al-Kbi" firstAttribute="leading" secondItem="Xxm-Vz-9PF" secondAttribute="leading" constant="16" id="m6J-bJ-Shc"/>
-                    <constraint firstItem="a81-Al-Kbi" firstAttribute="centerY" secondItem="Xxm-Vz-9PF" secondAttribute="centerY" id="waG-MJ-XaS"/>
-                </constraints>
-            </collectionViewCellContentView>
-            <connections>
-                <outlet property="logoImageView" destination="RHB-l7-KpO" id="89c-0v-piF"/>
-                <outlet property="titleStringLabel" destination="a81-Al-Kbi" id="Zgm-Er-OeU"/>
-            </connections>
-            <point key="canvasLocation" x="-157" y="25"/>
-        </collectionViewCell>
-    </objects>
-</document>
