--- Projects/UsertInterfaces/DesignSystem/Resources/RecommendPlayList/RecommendPlayListView.xib
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21507" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_12" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="RecommendPlayListView" customModule="DesignSystem" customModuleProvider="target">
-            <connections>
-                <outlet property="collectionView" destination="E2F-5w-GLv" id="edB-Ra-Pff"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB">
-            <rect key="frame" x="0.0" y="0.0" width="393" height="200"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="왁뮤팀이 추천하는 플레이리스트" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="jsU-pv-g93">
-                    <rect key="frame" x="20" y="32" width="353" height="24"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="24" id="z2f-LE-tgY"/>
-                    </constraints>
-                    <fontDescription key="fontDescription" type="boldSystem" pointSize="16"/>
-                    <nil key="textColor"/>
-                    <nil key="highlightedColor"/>
-                </label>
-                <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="E2F-5w-GLv">
-                    <rect key="frame" x="20" y="76" width="353" height="104"/>
-                    <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                    <collectionViewFlowLayout key="collectionViewLayout" minimumLineSpacing="10" minimumInteritemSpacing="10" id="AaT-1q-5ku">
-                        <size key="itemSize" width="128" height="128"/>
-                        <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                        <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                        <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
-                    </collectionViewFlowLayout>
-                </collectionView>
-            </subviews>
-            <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
-            <color key="backgroundColor" red="0.94117647058823528" green="0.95294117647058818" blue="0.96470588235294119" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-            <constraints>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="E2F-5w-GLv" secondAttribute="trailing" constant="20" id="1Rs-Am-Rim"/>
-                <constraint firstItem="E2F-5w-GLv" firstAttribute="top" secondItem="jsU-pv-g93" secondAttribute="bottom" constant="20" id="6Dh-qx-3ra"/>
-                <constraint firstItem="jsU-pv-g93" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="20" id="Tfd-Hv-iAf"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="trailing" secondItem="jsU-pv-g93" secondAttribute="trailing" constant="20" id="XZ3-sQ-FfP"/>
-                <constraint firstItem="vUN-kp-3ea" firstAttribute="bottom" secondItem="E2F-5w-GLv" secondAttribute="bottom" constant="20" id="fCe-hr-E9b"/>
-                <constraint firstItem="jsU-pv-g93" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="32" id="wKh-iW-JEb"/>
-                <constraint firstItem="E2F-5w-GLv" firstAttribute="leading" secondItem="vUN-kp-3ea" secondAttribute="leading" constant="20" id="zY4-N1-307"/>
-            </constraints>
-            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
-            <point key="canvasLocation" x="138" y="21"/>
-        </view>
-    </objects>
-</document>
