--- Projects/UsertInterfaces/DesignSystem/Resources/SongList/SongListCell.xib
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21225" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_0" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21207"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="SongListCellView" customModule="DesignSystem" customModuleProvider="target"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="default" indentationWidth="10" id="BbN-eK-qei" customClass="SongListCell" customModule="DesignSystem" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="390" height="44"/>
-            <autoresizingMask key="autoresizingMask"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" tableViewCell="BbN-eK-qei" id="Z7i-fu-ypt">
-                <rect key="frame" x="0.0" y="0.0" width="390" height="44"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFill" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="dummy_thumbnail_large" translatesAutoresizingMaskIntoConstraints="NO" id="QLV-Nm-lsR">
-                        <rect key="frame" x="20" y="2" width="72" height="40"/>
-                        <constraints>
-                            <constraint firstAttribute="width" constant="72" id="2Fd-XR-yxR"/>
-                            <constraint firstAttribute="height" constant="40" id="RO5-qW-mCX"/>
-                        </constraints>
-                    </imageView>
-                    <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" alignment="top" translatesAutoresizingMaskIntoConstraints="NO" id="exC-I4-ikk">
-                        <rect key="frame" x="99.999999999999986" y="1.6666666666666679" width="208.66666666666663" height="40.666666666666657"/>
-                        <subviews>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="nX7-Vj-Tit">
-                                <rect key="frame" x="0.0" y="0.0" width="41.333333333333336" height="20.333333333333332"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                                <nil key="textColor"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="WA7-bk-4w8">
-                                <rect key="frame" x="0.0" y="20.333333333333336" width="41.333333333333336" height="20.333333333333336"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                                <nil key="textColor"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                        </subviews>
-                    </stackView>
-                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="DuU-PZ-7D8">
-                        <rect key="frame" x="328.66666666666669" y="11.999999999999998" width="41.333333333333314" height="20.333333333333329"/>
-                        <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                        <nil key="textColor"/>
-                        <nil key="highlightedColor"/>
-                    </label>
-                </subviews>
-                <constraints>
-                    <constraint firstItem="exC-I4-ikk" firstAttribute="leading" secondItem="QLV-Nm-lsR" secondAttribute="trailing" constant="8" id="0T8-zP-KOz"/>
-                    <constraint firstItem="exC-I4-ikk" firstAttribute="centerY" secondItem="QLV-Nm-lsR" secondAttribute="centerY" id="BrL-fu-4Kq"/>
-                    <constraint firstItem="DuU-PZ-7D8" firstAttribute="leading" secondItem="exC-I4-ikk" secondAttribute="trailing" constant="20" id="QBb-Mk-xwn"/>
-                    <constraint firstItem="DuU-PZ-7D8" firstAttribute="centerY" secondItem="Z7i-fu-ypt" secondAttribute="centerY" id="VR6-f7-cFt"/>
-                    <constraint firstItem="QLV-Nm-lsR" firstAttribute="centerY" secondItem="Z7i-fu-ypt" secondAttribute="centerY" id="fYD-Kq-kRb"/>
-                    <constraint firstItem="QLV-Nm-lsR" firstAttribute="leading" secondItem="Z7i-fu-ypt" secondAttribute="leading" constant="20" id="ixf-SS-LlB"/>
-                    <constraint firstAttribute="trailing" secondItem="DuU-PZ-7D8" secondAttribute="trailing" constant="20" id="w7G-Ad-ery"/>
-                </constraints>
-            </tableViewCellContentView>
-            <connections>
-                <outlet property="albumImageView" destination="QLV-Nm-lsR" id="TrI-2N-E03"/>
-                <outlet property="artistLabel" destination="WA7-bk-4w8" id="b3R-gT-PZU"/>
-                <outlet property="releaseDateLabel" destination="DuU-PZ-7D8" id="nBL-qI-D3S"/>
-                <outlet property="titleLabel" destination="nX7-Vj-Tit" id="6gF-pa-ueC"/>
-            </connections>
-            <point key="canvasLocation" x="185" y="-163"/>
-        </tableViewCell>
-    </objects>
-    <resources>
-        <image name="dummy_thumbnail_large" width="324" height="182"/>
-    </resources>
-</document>
