--- Projects/UsertInterfaces/DesignSystem/Resources/Warning/WarningView.xib
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21225" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_0" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21207"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="System colors in document resources" minToolsVersion="11.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="WarningView" customModule="DesignSystem" customModuleProvider="target">
-            <connections>
-                <outlet property="superView" destination="Wjr-10-dsX" id="DyW-va-LFA"/>
-                <outlet property="warningImageView" destination="AtP-Ax-N0P" id="FWO-Uw-KLX"/>
-                <outlet property="warningLabelView" destination="GlZ-fm-vMf" id="7AM-Di-eJO"/>
-            </connections>
-        </placeholder>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <view contentMode="scaleToFill" id="iN0-l3-epB">
-            <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
-            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-            <subviews>
-                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Wjr-10-dsX">
-                    <rect key="frame" x="167" y="283" width="41.333333333333343" height="101"/>
-                    <subviews>
-                        <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="warning" translatesAutoresizingMaskIntoConstraints="NO" id="AtP-Ax-N0P">
-                            <rect key="frame" x="-19.333333333333343" y="0.0" width="80" height="80"/>
-                        </imageView>
-                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="GlZ-fm-vMf">
-                            <rect key="frame" x="0.0" y="80" width="41.333333333333336" height="21"/>
-                            <fontDescription key="fontDescription" type="system" pointSize="17"/>
-                            <nil key="textColor"/>
-                            <nil key="highlightedColor"/>
-                        </label>
-                    </subviews>
-                    <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-                    <constraints>
-                        <constraint firstAttribute="trailing" secondItem="GlZ-fm-vMf" secondAttribute="trailing" id="GlV-4D-Crb"/>
-                        <constraint firstAttribute="bottom" secondItem="GlZ-fm-vMf" secondAttribute="bottom" id="JEq-5q-qBN"/>
-                        <constraint firstItem="AtP-Ax-N0P" firstAttribute="top" secondItem="Wjr-10-dsX" secondAttribute="top" id="V7W-qX-UTF"/>
-                        <constraint firstItem="GlZ-fm-vMf" firstAttribute="leading" secondItem="Wjr-10-dsX" secondAttribute="leading" id="gF8-Bw-3dQ"/>
-                        <constraint firstItem="GlZ-fm-vMf" firstAttribute="top" secondItem="AtP-Ax-N0P" secondAttribute="bottom" id="riT-KR-8tF"/>
-                        <constraint firstItem="AtP-Ax-N0P" firstAttribute="centerX" secondItem="Wjr-10-dsX" secondAttribute="centerX" id="trg-X2-Faf"/>
-                    </constraints>
-                </view>
-            </subviews>
-            <viewLayoutGuide key="safeArea" id="vUN-kp-3ea"/>
-            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
-            <constraints>
-                <constraint firstItem="Wjr-10-dsX" firstAttribute="centerX" secondItem="iN0-l3-epB" secondAttribute="centerX" id="5ec-CB-SiD"/>
-                <constraint firstItem="Wjr-10-dsX" firstAttribute="centerY" secondItem="iN0-l3-epB" secondAttribute="centerY" id="7hj-59-75x"/>
-            </constraints>
-            <point key="canvasLocation" x="178" y="36"/>
-        </view>
-    </objects>
-    <resources>
-        <image name="warning" width="80" height="80"/>
-        <systemColor name="systemBackgroundColor">
-            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-        </systemColor>
-    </resources>
-</document>
